requirements.txt
flask
pdfplumber
docx2txt
groq
python-dotenv
werkzeug
os
json
